# optimize sampler cumsum