--- a/vllm/model_executor/layers/sampler.py
+++ b/vllm/model_executor/layers/sampler.py
@@ -3,6 +3,7 @@ from typing import Dict, List, Optional, Tuple
 
 import torch
 import torch.nn as nn
+import torch.nn.functional as F
 
 from vllm.model_executor.parallel_utils.communication_op import (
     tensor_model_parallel_gather)
@@ -184,6 +185,33 @@ def _apply_penalties(logits: torch.Tensor, prompt_tokens_tensor: torch.Tensor,
     logits -= presence_penalties.unsqueeze_(dim=1) * output_mask
     return logits
 
+def _cal_probs_sum(
+    probs_sort: torch.Tensor,
+    matmul_size: int = 1024
+):
+    """ Use matmul to speed up cumsum.
+
+    1. Pad probs_sort to make sure its last dim can be divisible by matmul_size.
+    2. Use upper triangular matrix multiplication to calculate the prefix sum in blocks.
+    3. Restore probs_sort and probs_sum to the original size.
+
+    Note that you can modify matmul_size to optimize matrix calculation time according to your GPU.
+    """
+    probs_size = probs_sort.size()
+    vocab_size = probs_size[-1]
+    # Pad probs_sort
+    padded_len = (vocab_size + matmul_size - 1) // matmul_size * matmul_size - vocab_size
+    probs_sort = F.pad(probs_sort, (0, padded_len))
+    cumsum_partition_size = probs_sort.size(-1) // matmul_size
+    # Use matmul to accelerate cumsum
+    triu_matrix = torch.triu(torch.ones(matmul_size, matmul_size, dtype=probs_sort.dtype, device=probs_sort.device))
+    probs_sum = torch.matmul(probs_sort.view(*probs_size[:-1], cumsum_partition_size, matmul_size), triu_matrix)
+    tmp = probs_sum[..., :-1, -1].cumsum(-1)
+    probs_sum[..., 1:, :].add_(tmp.unsqueeze(-1))
+    # Restore probs_sort and probs_sum
+    probs_sort = probs_sort[..., :vocab_size]
+    probs_sum = probs_sum.view(*probs_size[:-1], cumsum_partition_size * matmul_size)[..., :vocab_size]
+    return probs_sum
 
 def _apply_top_p_top_k(
     logits: torch.Tensor,
@@ -194,7 +222,8 @@ def _apply_top_p_top_k(
 
     # Apply top-p.
     probs_sort = logits_sort.softmax(dim=-1)
-    probs_sum = probs_sort.cumsum(dim=-1).sub_(probs_sort)
+    probs_sum = _cal_probs_sum(probs_sort)
+    probs_sum.sub_(probs_sort)
     top_p_mask = probs_sum > p.unsqueeze_(dim=1)
 
     # Apply top-k.