Speeding up cumsum in top_p sampling with upper triangular matrix multiplication. The idea originally comes from the 13th item of [https://zhuanlan.zhihu.com/p/673671771](url).

As we know, torch.cumsum uses cuda core, while matmul uses tensor core. Typically, tensor core is much faster than cuda core. Therefore, we use upper triangular matrix multiplication to calculate the cumsum in top_p sampling. This change decreases top_p sampling time from 6.8ms to 1.4ms for Llama-13B, whose vocab size is 32000. Here are our test settings:
- GPU = 1 A100-80G
- model = Llama-13B
- vocab size = 32000
- vllm = 0.2.7
- top_p = 0.80
- input length = 300
- output length = 200

### Nsys Perf
#### Before Optimization
![优化前](https://github.com/vllm-project/vllm/assets/44199326/543cf3e2-5f5b-4ac6-b61d-0fe006d06231)
#### After Optimization
![优化后](https://github.com/vllm-project/vllm/assets/44199326/79666601-f92b-4049-9a1e-4e19ab4ac2fa)

### Latency
| QPS | Latency before opt (s) | Latency after opt (s) | Latency decrease (%) |
|--|--|--|--|
| 1 | 4.9314 s | 4.6609 s | 5.48% |
| 2 | 5.7945 s | 5.4266 s | 6.34% |
| 4 | 7.9615 s | 7.5823 s | 4.76% |