@AniZpZ 2024-01-24
> @HandH1998 would it be possible to add a unit test to ensure consistency between your implementation and torch? I would suggest generating random data for like 256 set seeds and asserting that results are all close.

We are pleased to add a unit test, using upper triangular matmul to calculate prefix sum is mathematically equivalent to cumsum.