@HandH1998 2024-01-24
@Yard1, hi, we have added a unit test to ensure consistency between two implementations. There are two points that need to be explained:
1. When the `dtype` is `torch.float32`, the two implementations are always consistent.
2. When the `dtype` is `torch.float16` and `vocab_size` is large (>=5000), they give different results. We think it is because that there are more overflows for `torch.float16` in our unit test. As the probability distribution in practical LLM sampling is much sharper (the top-k tokens may occupy the main probability), we think the new implementation can still work in practice.