@HandH1998 2024-01-31
@Yard1, we have merged upstream to our branch. But there is a case that failed when running `tests/samplers/test_sampler.py::test_sampler_top_k_top_p` for 'seed =  93`. 
![image](https://github.com/vllm-project/vllm/assets/44199326/fa0d3df0-c18b-4e35-8a8d-0642d10350a4)
However, it can pass all the cases of the unit test in our machine (as the following picture). It is a little weird. Hope you can check it. 
<img width="1364" alt="sampler" src="https://github.com/vllm-project/vllm/assets/44199326/2ebbda2b-3bb7-4765-b158-ba81040d834d">