@HandH1998 2024-02-19
> @HandH1998 Sorry, let me take a look at this soon! We may just need to relax the tolerances

Hi, have you relaxed the tolerances? Could you please give us any other advice?