@Yard1 2024-03-01
@HandH1998 As long as the fp32 test is passing, it should be fine. What I would suggest is keeping a tighter tolerance for fp32 and relaxing the tolerance for fp16.