@HandH1998 2024-03-04
@Yard1 Hello, thanks for your reply! I have two points to discuss. 
1. The unit test failure is at `test_sampler.py::test_sampler_top_p_top_k`(as you can see in [https://github.com/vllm-project/vllm/pull/2510#issuecomment-1918470080](https://github.com/vllm-project/vllm/pull/2510#issuecomment-1918470080)) which is not our new added unit test funciton. We don't think it should be modified by us. Thus, we hope you can relax the tolerance for `test_sampler.py::test_sampler_top_p_top_k` in github cr.

2. As for our unit test function `test_sampler.py::test_optimized_cumsum`, I will try to do as what you suggest, i.e.,  keeping a tighter tolerance for fp32 and relaxing the tolerance for fp16.