@Yard1 2024-03-05
Looking at the specifics of the failure in the `test_sampler_top_p_top_k[125]` suggests a bug:

```
(Pdb) hf_probs.max(dim=-1)
torch.return_types.max(
values=tensor([1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.,
        1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.,
        1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.,
        1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.,
        1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.,
        1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.,
        1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.],
       device='cuda:0'),
indices=tensor([28559, 27816, 20757, 24047,  3210, 24129, 11583,  1129, 22047,  5554,
        26774, 27723,  8267,  6854, 15496, 10581, 17418, 23408, 25461, 18058,
        11906,  4824, 31754,  9747, 11349, 17897, 11644, 12138, 20257,  6374,
        27749,  4806, 14131, 13845,  1245, 10751,  2044, 11240, 19818, 14420,
        28000, 22014, 31369, 28155, 22590,  1882, 20337, 17083, 24837, 28885,
        10601, 21028,  1775, 18500, 24188,  9904, 21428,  4079, 20943, 30265,
         1242,  5908,  1494, 12515, 18015, 22230, 14465, 20647, 27481, 15786,
        30594, 30998,  5197, 18158, 16708, 16184, 20515,  9649, 11981, 24253,
         1758, 17008, 25068,   798, 31775, 15536, 30979, 21243, 12164, 28002,
        24322, 23199, 21545, 29223,  3806,  8886,  2921, 27718, 17745,  7387,
        29318, 15423,  9530, 11149, 29659,  3633, 19103, 11518,  9710,  5787,
         7602, 10742, 22062, 18175, 15502,  7052, 18804, 16051, 14226,  4399,
          582,   581, 22743, 28021, 26454], device='cuda:0'))
(Pdb) sample_probs.max(dim=-1)
torch.return_types.max(
values=tensor([1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000,
        1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000,
        1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000,
        1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000,
        1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 0.5117, 1.0000, 1.0000,
        1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000,
        1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000,
        1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000,
        1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000,
        1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000,
        1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000,
        1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000,
        1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000,
        1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000, 1.0000],
       device='cuda:0'),
indices=tensor([28559, 27816, 20757, 24047,  3210, 24129, 11583,  1129, 22047,  5554,
        26774, 27723,  8267,  6854, 15496, 10581, 17418, 23408, 25461, 18058,
        11906,  4824, 31754,  9747, 11349, 17897, 11644, 12138, 20257,  6374,
        27749,  4806, 14131, 13845,  1245, 10751,  2044, 11240, 19818, 14420,
        28000, 22014, 31369, 28155, 22590,  1882, 20337, 17083, 24837, 28885,
        10601, 21028,  1775, 18500, 24188,  9904, 21428,  4079, 20943, 30265,
         1242,  5908,  1494, 12515, 18015, 22230, 14465, 20647, 27481, 15786,
        30594, 30998,  5197, 18158, 16708, 16184, 20515,  9649, 11981, 24253,
         1758, 17008, 25068,   798, 31775, 15536, 30979, 21243, 12164, 28002,
        24322, 23199, 21545, 29223,  3806,  8886,  2921, 27718, 17745,  7387,
        29318, 15423,  9530, 11149, 29659,  3633, 19103, 11518,  9710,  5787,
         7602, 10742, 22062, 18175, 15502,  7052, 18804, 16051, 14226,  4399,
          582,   581, 22743, 28021, 26454], device='cuda:0'))
```

Notice the `0.5117`. I wouldn't be comfortable merging with such a huge divergence.