@HandH1998 2024-03-06
@Yard1 Hi, thanks for your test! I was wondering if this test was done on your machine or github cr env? In my test, both `hf_probs` and `sample_probs` have something like `0.5117` which is not equal to 1.0, but there is no such a huge divergence like what you found. As a result, all the cases of `test_sampler_top_p_top_k` can pass the test in my machine. It is so weird. I still need time to figure it.