@Yard1 2024-03-06
@HandH1998 I ran this on an A10 GPU