@HandH1998 2024-03-06
@Yard1 I fixed the softmax dtype to `torch.float` in `_apply_top_k_top_p`([https://github.com/vllm-project/vllm/pull/2510/commits/f9b945521cd99b8542d83343fdcb4007c2e16091](https://github.com/vllm-project/vllm/pull/2510/commits/f9b945521cd99b8542d83343fdcb4007c2e16091)). It can work now. I wonder if the change is accepted.